001-001 Dynamic display using streambuilder
