+++ lib/main.dart
@@ -2,10 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
 import 'package:firebase_core/firebase_core.dart';
-import 'package:sample_flutter_app/model/talk_room.dart';
 
-import 'app.dart';
-import 'model/app_state_model.dart';
+import 'screens/app.dart';
+import 'models/app_state_model.dart';
 
 Future<void> main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -14,7 +13,6 @@ Future<void> main() async {
     MultiProvider(
       providers: [
         ChangeNotifierProvider<AppStateModel>(create: (_) => AppStateModel()..loadProducts()),
-        ChangeNotifierProvider<TalkProvider>(create: (_) => TalkProvider()..fetchTalkData()),
       ],
       child: BaseApp(),
     ),
