+++ lib/models/app_state_model.dart
