+++ lib/screens/home_tab.dart
@@ -1,8 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
-
-import 'model/app_state_model.dart';
-import 'product_row_item.dart';
+import 'package:sample_flutter_app/components/product_row_item.dart';
+import 'package:sample_flutter_app/models/app_state_model.dart';
 
 class HomeTab extends StatelessWidget {
   @override
