+++ lib/screens/time_line.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/cupertino.dart';
 import 'package:provider/provider.dart';
-
-import 'model/app_state_model.dart';
+import 'package:sample_flutter_app/models/app_state_model.dart';
 
 class TimeLineTab extends StatefulWidget {
   @override
